// kitchen-planner-fe-main/src/app/api/api-kms-planner-masterdata/model/deliveryAddress.ts
/**
 * KMS - Kitchen Management System API
 *
 * Contact: anas-ur-rasheed.abid@tu-ilmenau.de
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


export interface DeliveryAddress { 
    /**
     * Street address
     */
    street: string;
    /**
     * City name
     */
    city: string;
    /**
     * State or province
     */
    state: string;
    /**
     * ZIP or postal code
     */
    zipCode: string;
    /**
     * Country name
     */
    country: string;
    /**
     * Special delivery instructions (optional)
     */
    instructions?: string;
}