// kitchen-planner-fe-main/src/app/api/api-kms-planner-masterdata/model/orderCreate.ts
/**
 * KMS - Kitchen Management System API
 *
 * Contact: anas-ur-rasheed.abid@tu-ilmenau.de
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */
import { DeliveryAddress } from './deliveryAddress';
import { BillingInfo } from './billingInfo';
import { CheckoutRequestArrayOfMealsInner } from './checkoutRequestArrayOfMealsInner';


export interface OrderCreate { 
    /**
     * Array of meals with quantities
     */
    meals: Array<CheckoutRequestArrayOfMealsInner>;
    /**
     * Total order amount including all fees and taxes
     */
    total: number;
    /**
     * Subtotal before fees and taxes
     */
    subTotal: number;
    /**
     * Tax amount
     */
    tax: number;
    /**
     * Delivery fee
     */
    deliveryFee: number;
    /**
     * Discount amount
     */
    discount: number;
    billingInfo: BillingInfo;
    deliveryAddress: DeliveryAddress;
    /**
     * Special instructions for the order
     */
    specialInstructions?: string;
}